move existing wdk-macro tests to integration tests folder
